+++ rest/conversations/v1/model_conversations_v1_conversation_message.go
@@ -48,4 +48,6 @@ type ConversationsV1ConversationMessage struct {
 	Delivery *interface{} `json:"delivery,omitempty"`
 	// Absolute URL to access the receipts of this message.
 	Links *map[string]interface{} `json:"links,omitempty"`
+	// The unique ID of the multi-channel Rich Content template.
+	ContentSid *string `json:"content_sid,omitempty"`
 }